+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleazed]
+
+### Added
+
+- `no-invalid-cosmetic-separator` linter rule.
+
 ## [4.0.0-beta.5] - 2025-12-22
 
 ### Added